refactor: move browser tools under tools
