+++ packages/browseros-agent/apps/server/src/agent/chat-mode.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import { BROWSER_TOOLS } from '../browser-tools/registry'
+import { BROWSER_TOOLS } from '../tools/browser/registry'
 
 export const CHAT_MODE_ALLOWED_TOOLS = new Set([
   ...BROWSER_TOOLS.filter((tool) => tool.annotations?.readOnlyHint).map(
