+++ packages/browseros-agent/apps/server/src/tools/browser/act.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import type { BrowserSession } from '../browser/core/session'
+import type { BrowserSession } from '../../browser/core/session'
 import {
   defineTool,
   errorResult,
