+++ packages/browseros-agent/apps/server/src/tools/browser/registry.ts
