+++ packages/browseros-agent/apps/server/src/tools/browser/screenshot.ts
