+++ packages/browseros-agent/apps/server/src/tools/browser/snapshot.ts
